requirements.txt
# RAG系统依赖包
gensim==4.3.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.35.0
torch==2.1.0
dashscope==1.17.0
jieba==0.42.1
tqdm==4.66.1